do not run TestScanNullTime on Go1.13+
