+++ nulltime_test.go
@@ -1,62 +1,43 @@
-// Go MySQL Driver - A MySQL-Driver for Go's database/sql package
-//
-// Copyright 2013 The Go-MySQL-Driver Authors. All rights reserved.
-//
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// You can obtain one at http://mozilla.org/MPL/2.0/.
-
 package mysql
 
 import (
-	"database/sql"
-	"database/sql/driver"
 	"testing"
 	"time"
 )
 
-var (
-	// Check implementation of interfaces
-	_ driver.Valuer = NullTime{}
-	_ sql.Scanner   = (*NullTime)(nil)
-)
+func TestNullTime(t *testing.T) {
+	runTests(t, dsn+"&parseTime=true&loc=US%2FCentral", func(dbt *DBTest) {
+		// Create table
+		dbt.mustExec("CREATE TABLE test (ts TIMESTAMP)")
 
-func TestScanNullTime(t *testing.T) {
-	var scanTests = []struct {
-		in    interface{}
-		error bool
-		valid bool
-		time  time.Time
-	}{
-		{tDate, false, true, tDate},
-		{sDate, false, true, tDate},
-		{[]byte(sDate), false, true, tDate},
-		{tDateTime, false, true, tDateTime},
-		{sDateTime, false, true, tDateTime},
-		{[]byte(sDateTime), false, true, tDateTime},
-		{tDate0, false, true, tDate0},
-		{sDate0, false, true, tDate0},
-		{[]byte(sDate0), false, true, tDate0},
-		{sDateTime0, false, true, tDate0},
-		{[]byte(sDateTime0), false, true, tDate0},
-		{"", true, false, tDate0},
-		{"1234", true, false, tDate0},
-		{0, true, false, tDate0},
-	}
+		// Insert local time into database (should be converted)
+		usCentral, _ := time.LoadLocation("US/Central")
+		reftime := time.Date(2014, 05, 30, 18, 03, 17, 0, time.UTC).In(usCentral)
+		dbt.mustExec("INSERT INTO test VALUE (?)", reftime)
 
-	var nt = NullTime{}
-	var err error
+		// Retrieve time from DB
+		rows := dbt.mustQuery("SELECT ts FROM test")
+		defer rows.Close()
+		if !rows.Next() {
+			dbt.Fatal("did not get any rows out")
+		}
 
-	for _, tst := range scanTests {
-		err = nt.Scan(tst.in)
-		if (err != nil) != tst.error {
-			t.Errorf("%v: expected error status %t, got %t", tst.in, tst.error, (err != nil))
+		var dbTime NullTime
+		err := rows.Scan(&dbTime)
+		if err != nil {
+			dbt.Fatal("Err", err)
 		}
-		if nt.Valid != tst.valid {
-			t.Errorf("%v: expected valid status %t, got %t", tst.in, tst.valid, nt.Valid)
+
+		// Check that dates match
+		if reftime.Unix() != dbTime.Time.Unix() {
+			dbt.Errorf("times do not match.\n")
+			dbt.Errorf(" Now(%v)=%v\n", usCentral, reftime)
+			dbt.Errorf(" Now(UTC)=%v\n", dbTime)
 		}
-		if nt.Time != tst.time {
-			t.Errorf("%v: expected time %v, got %v", tst.in, tst.time, nt.Time)
+		if dbTime.Time.Location().String() != usCentral.String() {
+			dbt.Errorf("location do not match.\n")
+			dbt.Errorf(" got=%v\n", dbTime.Time.Location())
+			dbt.Errorf(" want=%v\n", usCentral)
 		}
-	}
+	})
 }
